+++ neuron/src/app/Neuron/CLI/New.hs
@@ -55,7 +55,7 @@ newZettelFile NewCommand {..} config = do
       liftIO $ do
         fileAction :: FilePath -> FilePath -> IO () <-
           bool (pure showAction) mkEditActionFromEnv edit
-        writeFileText (notesDir </> zettelFile) $ defaultZettelContent zettelFormat date title
+        writeFileText (notesDir </> zettelFile) $ defaultZettelContent zettelFormat date
         fileAction notesDir zettelFile
   where
     mkEditActionFromEnv :: IO (FilePath -> FilePath -> IO ())
@@ -82,8 +82,8 @@ newZettelFile NewCommand {..} config = do
           if null v then pure Nothing else pure (Just v)
 
 -- TODO use configurable template files?
-defaultZettelContent :: ZettelFormat -> DateMayTime -> Maybe Text -> Text
-defaultZettelContent format (formatDateMayTime -> date) mtitle = case format of
+defaultZettelContent :: ZettelFormat -> DateMayTime -> Text
+defaultZettelContent format (formatDateMayTime -> date) = case format of
   ZettelFormat_Markdown ->
     T.intercalate
       "\n"
@@ -91,18 +91,15 @@ defaultZettelContent format (formatDateMayTime -> date) mtitle = case format of
         "date: " <> date,
         "---",
         "",
-        "# " <> title,
+        "# " <> "Untitled",
         "\n"
       ]
   ZettelFormat_Org ->
     T.intercalate
       "\n"
-      [ "* " <> title,
+      [ "* " <> "Untitled",
         "    :PROPERTIES:",
         "    :Date: " <> date,
         "    :END:",
         "\n"
       ]
-  where
-    defaultTitleName = "Zettel created on " <> date
-    title = maybe defaultTitleName T.strip mtitle